blog agent/requirements.txt
phidata
python-dotenv
beautifulsoup4
packaging
fastapi
uvicorn
groq
requests
google-search-results